requirements.txt
asgiref==3.6.0
Django==4.1.6
django-autoslug==1.9.8
django-robokassa-payments==1.3.3
Pillow==9.4.0
product==0.1.1.1
pyaes==1.6.1
pyasn1==0.4.8
rsa==4.9
sqlparse==0.4.3
